requirements-dev.txt
coverage==3.6
django-debug-toolbar==0.9.4
django-debug-toolbar-template-timings==0.3
django-discover-runner==0.3
django-webtest==1.6
factory-boy==1.2.0
pep8==1.4.5
PyQuery==1.2.4
WebTest==2.0.1
flake8==2.0
